+++ MapboxNavigation/ImageRepository.swift
@@ -1,14 +1,5 @@
 import Foundation
 
-typealias NoArgBlock = () -> Void
-
-protocol BimodalImageCache {
-    func store(_ image: UIImage, forKey key: String, toDisk: Bool, completion completionBlock: NoArgBlock?)
-    func imageFromCache(forKey: String?) -> UIImage?
-    func clearMemory()
-    func clearDisk(completion: NoArgBlock?)
-}
-
 class ImageRepository {
 
     public var sessionConfiguration: URLSessionConfiguration = URLSessionConfiguration.default {
@@ -29,7 +20,7 @@ class ImageRepository {
         imageCache = cache
     }
 
-    func resetImageCache(_ completion: NoArgBlock?) {
+    func resetImageCache(_ completion: CompletionHandler?) {
         imageCache.clearMemory()
         imageCache.clearDisk(completion: completion)
     }
@@ -39,7 +30,7 @@ class ImageRepository {
     }
 
     func cachedImageForKey(_ key: String) -> UIImage? {
-        return imageCache.imageFromCache(forKey: key)
+        return imageCache.image(forKey: key)
     }
 
     func imageWithURL(_ imageURL: URL, cacheKey: String, completion: @escaping (UIImage?) -> Void) {